Move Plane2D into Geometry namespace
