--- Assets/Tests/Data Structures/Plane2D_Tests.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: ba70d64cc0fe1d44f82eabf3066e7626
